pyproject.toml
[project]
name = "aeromind-dag-architect"
version = "0.1.0"
description = "An intelligent AI agent system for automatically generating, deploying, and testing Apache Airflow DAGs from natural language requirements"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.7.0,<3.0.0",
    "apache-airflow-providers-docker>=3.9.0,<4.0.0",
    "apache-airflow-providers-postgres>=5.7.0,<6.0.0",
    "crewai>=0.28.0",
    "docker>=7.0.0,<8.0.0",
    "google-generativeai>=0.4.1,<0.5.0",
    "ipython>=8.0.0,<9.0.0",
    "langchain>=0.1.0,<0.2.0",
    "langchain-community>=0.0.20",
    "langchain-google-genai>=0.0.10,<0.1.0",
    "langchain-huggingface==0.1.0",
    "langchain-openai>=0.0.5,<0.1.0",
    "langgraph>=0.0.15",
    "numpy>=1.24.0,<2.0.0",
    "openai>=1.0.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pytest>=7.4.0,<9.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "requests>=2.28.0,<3.0.0",
    "sentence-transformers>=2.2.0,<3.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "torch>=2.0.0,<3.0.0",
    "tqdm>=4.65.0,<5.0.0",
    "transformers>=4.30.0,<5.0.0",
]
